+++ src/app.ts
@@ -1,6 +1,5 @@
 import { MemorySaver, InMemoryStore, Command } from '@langchain/langgraph';
 import { HumanMessage } from '@langchain/core/messages';
-import * as core from '@actions/core';
 import { StateGraph } from '@langchain/langgraph';
 import { AIMessage } from '@langchain/core/messages';
 import { isAIMessage } from '@langchain/core/messages';
@@ -32,12 +31,20 @@ import {
   finalNotesAgent
 } from './agents';
 
-export const MainGraphRun = async (): Promise<string> => {
+export const MainGraphRun = async ({
+  fileName,
+  recursionLimit,
+  additionalPrompt
+}: {
+  fileName: string;
+  recursionLimit: number;
+  additionalPrompt: string;
+}): Promise<string> => {
   // Initialize memory to persist state between graph runs
   const checkpointer = new MemorySaver();
   const inMemoryStore = new InMemoryStore();
 
-  const filename: string = core.getInput('file_name');
+  const filename: string = fileName;
   const toolNames = CustomTools.map(tool => tool.name).join(', ');
 
   const toolExecutor = async (state: State) => {
@@ -180,6 +187,10 @@ export const MainGraphRun = async (): Promise<string> => {
   const app = workflow.compile({ checkpointer, store: inMemoryStore });
   console.log('app version', 'v0.1.54-alpha.10');
 
+  const additionalPromptNotes = `
+  Additional Notes: ${additionalPrompt}
+  `;
+
   const query = `
   You are a coding assistant with expertise in test automation.
   Generate and execute tests for ${filename}.
@@ -193,7 +204,13 @@ export const MainGraphRun = async (): Promise<string> => {
   3. Improve existing tests or create new tests
   4. Save test file
   5. Run tests with coverage
-  6. Fix any failures`;
+  6. Fix any failures
+  7. Analyze test results
+  8. Provide final notes
+
+  ${additionalPrompt ? additionalPromptNotes : ''}
+
+  `;
 
   // Use the Runnable
   const currentDate = new Date().toISOString().replace('T', ' ').split('.')[0];
@@ -202,7 +219,7 @@ export const MainGraphRun = async (): Promise<string> => {
       messages: [new HumanMessage(query)],
       fileName: filename
     },
-    { recursionLimit: 100, configurable: { thread_id: 1001 } }
+    { recursionLimit: recursionLimit || 200, configurable: { thread_id: 1001 } }
   );
 
   const resultOfGraph = finalState.finalComments;